fix: resolve 4 critical bugs from PR review
- user-service/utils/jwt.py: fix datetime.UTC (AttributeError at runtime)
  use datetime.now(timezone.utc) instead of datetime.now(datetime.UTC)

- order-service/order_service.py: replace fragile 3-branch product price
  parsing with _extract_price() helper that raises on missing/invalid
  price instead of silently defaulting to 0

- order-service/order_service.py + clients.py: add stock rollback on
  payment failure. release_stock() calls POST /stock/increment as a
  compensating action; _rollback_stock() wraps it with error logging.
  stock_reserved flag prevents spurious rollbacks before decrement.

- order-service + payment-service: change amount fields from float/Float
  to Numeric(12,2)/Decimal across models, Pydantic contracts, and the
  charge_payment client to avoid floating-point rounding on money

+++ services/order-service/app/contracts/order.py
@@ -1,6 +1,7 @@
 """Order contracts."""
 
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
@@ -18,7 +19,7 @@ class OrderResponse(BaseModel):
     user_id: str
     product_id: str
     quantity: int
-    total_amount: float
+    total_amount: Decimal
     status: str
     payment_id: str | None = None
     idempotency_key: str
