+++ services/payment-service/app/contracts/payment.py
@@ -1,13 +1,14 @@
 """Payment contracts."""
 
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
 
 class ChargeRequest(BaseModel):
     order_id: str = Field(..., min_length=1)
-    amount: float = Field(..., gt=0)
+    amount: Decimal = Field(..., gt=0)
     currency: str = Field(default="USD", max_length=3)
     idempotency_key: str = Field(..., min_length=1, max_length=255)
     card_last_four: str | None = Field(None, max_length=4)
@@ -17,7 +18,7 @@ class PaymentResponse(BaseModel):
     id: str
     idempotency_key: str
     order_id: str
-    amount: float
+    amount: Decimal
     currency: str
     status: str
     card_last_four: str | None = None
